hardware_connection/requirements.txt
paho-mqtt
simple-rest-client